feat(loans, mutual-funds): Implement loan management and enhance fund details
This commit introduces a comprehensive loan management system, allowing users to view, add, edit, and delete loans. It also refactors pagination across the app into reusable components and refines the mutual fund details and bundle experiences.

**Key Changes:**

*   **Loan Management:**
    *   **New Screens:** Introduced `LoansScreen` (list view), `AddLoanScreen` (create/edit), and `LoanFlowScreen` (navigation container).
    *   **Loan Functionality:** Supports fetching paginated loans, viewing loan details by ID, and performing CRUD operations (Create, Read, Update, Delete).
    *   **ViewModels:** Added `LoanInfoViewModel` for the list state and `AddLoanViewModel` for form handling and submission logic.
    *   **Data Layer:** Implemented `UserLoanDto`, `SingleLoanDto`, and `UpdateLoanRequest`. Expanded `UserFinance` repository and implemented real API calls in `UserFinanceRepo`.
    *   **Use Cases:** Added `GetLoansUseCase`, `GetLoanByIdUseCase`, `AddSingleLoanUseCase`, `UpdateSingleLoanUseCase`, and `DeleteLoanUseCase`.

*   **Shared Pagination Components:**
    *   **`PaginationEffect.kt`:** A reusable `LaunchedEffect` wrapper that triggers a callback when the user scrolls near the end of a `LazyList`.
    *   **`PaginationFooter.kt`:** A standardized loading indicator for the bottom of paginated lists.
    *   **Integration:** Refactored `FDSearchResultScreen` and `MutualFundSearchResultRoot` to use these new shared components.

*   **Mutual Fund & Bundle Enhancements:**
    *   **Fund Details (`MutualFundDetailsScreen.kt`):** Improved the "NAV Change" display to dynamically calculate percentage changes based on the selected graph duration (1M, 3M, 1Y, etc.).
    *   **Bundle Search (`BundleResultScreenRoot.kt`):** Added a "Cart FAB" to the bundle results and all-bundles screens.
    *   **Investment Chips:** Introduced `generateInvestmentChips` utility to dynamically suggest investment amounts based on the fund's minimum requirement and whether it's a SIP or One-time investment.
    *   **Metrics:** Updated `Metrics.kt` DTO to handle nullable return fields safely.

*   **UI & Navigation:**
    *   **`AppNavigation.kt`:** Registered new loan routes and updated existing routes to support cart navigation from search results.
    *   **Home Screen:** `HomeScreenViewModel` now fetches user cart data on initialization.
    *   **Loan UI:** Updated `ExpandableLoanEntry` to support an optional edit action.
    *   **Fire Report:** Re-enabled the "Loan & Debts" update card in the FIRE report flow.

*   **Technical Refinements:**
    *   **App Events:** Added `LoanEventRefresh` to `AppEventsController` to trigger UI updates across the app when loans are modified.
    *   **Versioning:** Incremented `versionCode` to 10 and `versionName` to 1.2.1 in `build.gradle.kts`.
    *   **Formatting:** Fixed a bug in `FixedDespositCard.kt` where amount formatting failed due to type mismatch.

+++ composeApp/src/commonMain/kotlin/org/sharad/velvetinvestment/shared/compose/FixedDespositCard.kt
@@ -94,7 +94,7 @@ fun FixedDepositCard(fdData: FixedDepositPortfolioDomain, onClick: () -> Unit){
                 }
 
                 Text(
-                    text = formatMoneyAfterL(fdData.amount.toLong()),
+                    text = formatMoneyAfterL(fdData.amount.toDouble().toLong()),
                     style = MaterialTheme.typography.labelSmall,
                     color = darkBlue
                 )
